printing status at the end
SGs are not always being deleted in time, and since that info had scrolled off the screen, i was missing it. so this gathers all the errors up and dumps them at the end. the consolidation of the functions into a list and for loop is just so that we don't have to check for an error each time. This also creates an env var so that we can just run one of the funcs (e.g. delete_sg) if needed.

The plugin show change gets me around a hang that happens when i'm prompted about a new version.

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>