lookup dhcp id
so there's a way to do this without curl. we need to think about how to get us around the destroy chicken and egg problem that users can easily deploy. this is one part. we also need to be able to import the dhcp resource so it can be deleted via tf

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>